# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AccessCopilot API",
    "description": "AI-powered accessibility action engine backend.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/perception/analyze": {
      "post": {
        "tags": [
          "perception"
        ],
        "summary": "Analyze Frame",
        "operationId": "analyze_frame_api_v1_perception_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeFrameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeFrameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scene/update": {
      "post": {
        "tags": [
          "perception"
        ],
        "summary": "Update Scene",
        "description": "Apply detections to the stateful scene graph for a session.",
        "operationId": "update_scene_api_v1_scene_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assistant/query": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Assistant Query",
        "description": "Handle a user's voice/text command.\n\nThe LangGraph agent gets wired here in the agent milestone. Until\nthen a deterministic fallback provides graceful degraded responses.",
        "operationId": "assistant_query_api_v1_assistant_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Action": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "instruction"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.5
          },
          "haptics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Haptics"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "Action",
        "description": "The recommended next action — the core product output."
      },
      "AnalyzeFrameRequest": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "encoded_frame": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoded Frame"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "session_id"
        ],
        "title": "AnalyzeFrameRequest",
        "description": "A single frame sent for analysis (frames are throttled client-side)."
      },
      "AnalyzeFrameResponse": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Detections"
          },
          "scene_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scene Summary"
          },
          "processing_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Processing Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "detections"
        ],
        "title": "AnalyzeFrameResponse",
        "description": "Structured result; raw frames are never stored."
      },
      "AssistantQuery": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AssistantQuery",
        "description": "A voice-text command from the user (Whisper transcript)."
      },
      "AssistantResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/SceneObject"
            },
            "type": "array",
            "title": "Detections"
          },
          "route_change": {
            "type": "boolean",
            "title": "Route Change",
            "default": false
          },
          "debug": {
            "additionalProperties": true,
            "type": "object",
            "title": "Debug"
          }
        },
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "AssistantResponse",
        "description": "Natural-language response, always grounded in structured state."
      },
...
      "Detection": {
        "properties": {
          "object_type": {
            "type": "string",
            "title": "Object Type"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "estimated_distance_m": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Distance M"
          },
          "distance_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Confidence"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction",
            "default": "front"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "maxItems": 4,
                "minItems": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "detector"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "object_type",
          "confidence"
        ],
        "title": "Detection",
        "description": "A single perceived object, fused from detection + depth + OCR."
      },
      "Direction": {
        "type": "string",
        "enum": [
          "front",
          "left",
          "right",
          "behind"
        ],
        "title": "Direction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneObject": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction",
            "default": "front"
          },
          "accessible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessible"
          },
          "temporary_blockage": {
            "type": "boolean",
            "title": "Temporary Blockage",
            "default": false
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.5
          },
          "currently_visible": {
            "type": "boolean",
            "title": "Currently Visible",
            "default": true
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "SceneObject",
        "description": "A node in the scene graph derived from one or more detections."
      },
      "SceneUpdate": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Detections"
          },
          "location": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "SceneUpdate",
        "description": "Incremental update of the live scene graph."
      },
      "SceneUpdateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "scene_objects": {
            "items": {
              "$ref": "#/components/schemas/SceneObject"
            },
            "type": "array",
            "title": "Scene Objects"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "observation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "scene_objects",
          "updated_at"
        ],
        "title": "SceneUpdateResponse",
        "description": "The merged scene after applying the update."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}